python_app/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
lxml>=3.1.0
typing-extensions>=4.9.0
Pillow==10.1.0 